Remake the shopping cart store function, add filters and user doping selection
The implementation of the operation of adding products to the shopping cart was remake to find equal products. Automatic recalculation of prices when changing the size of pizza and change product doping was added. A product category filter was added. New tables were created and relationships between models were realized. Pop up window with product details was remaked. Error related with product catalog elements and ajax queries was fixed.