Add roles and permissions CRUD, admin dashboard, trait for permissions
Role and permission CRUD functional was added to the admin part. New migrations were applied to create the tables for storing the roles and permissions. A trait was created to determine the permissions of the current user. A new routes were added to the project. Ajax request for record removing was created.