Add a progress bar, add order history, fix ajax request errors
A progress bar was added to display the load status of the shopping cart. Controller and views were optimized with computed fields. A user info page was added to the shopping cart, payment page was added. The routes for new controllers were created. Errors related with ajax request were fixed.