Add the groups for select box in filters, fix errors in product catalog, fix routes, shopping cart
Categories and subcategories were added to the select box in filters (in the product catalog). Fix error related with product details info view. Shopping cart was fixed, add id parameter to the ajax request for correct operation. Self-reference relationship was added to the ProductCategory model. Errors in the money format viewing were fixed.